+++ lib/severance/application.ex
@@ -25,8 +25,8 @@ defmodule Severance.Application do
   end
 
   @spec dispatch(CLI.parse_args_result()) :: {:ok, pid()} | no_return()
-  defp dispatch(:help) do
-    IO.puts(CLI.usage())
+  defp dispatch({:help, path}) do
+    IO.puts(CLI.usage(path))
     System.halt(0)
   end